requirements.txt
streamlit
matplotlib
google-cloud-bigquery
vertexai
bcrypt